// src/app/precisa/proyecto/Programa/model/DtoPrograma.ts
export class DtoPrograma {
    constructor() {
    }
    CompaniaCodigo: string;
    DescripcionCorta:string;
    IdSede: number;
    IdPromotor:number;
    SedDescripcion:string;
    Departamento: string;
    Provincia: string;
    Distrito: string;
    IdProyecto: number;
    promotordesc:string;
    Sucursal: string;
    NIdPromotor:number;
    Codigo: string;
    Nombre: string;
    FechaRegistro:Date;
    Area: number;
    Precio: number;
    MonedaCodigo: string;
    DesMoneda: string;
    TotalCuota: number;
    Ubigeo: string;
    MontoTotal: number;
    MontoInicial: number;
    Estado: number;
    UsuarioCreacion: string;
    UsuarioModificacion: string;
    FechaCreacion:Date;
    FechaModificacion:Date;
    IpCreacion: string;
    IpModificacion: string;
    Estado1: string;
    TipoDocumento: string;
    Documento: string;
    NombreCompleto: string;
    ubigeodesc: string;

}